Hoist OwnedWorldCalculation to its own file.
Step 4 of the Big Break Up.
