+++ PyRoute/Galaxy.py
@@ -14,8 +14,8 @@ import math
 import networkx as nx
 
 from PyRoute.Star import Star
-from PyRoute.Calculation.TradeCalculation import TradeCalculation, CommCalculation, \
-    OwnedWorldCalculation
+from PyRoute.Calculation.TradeCalculation import TradeCalculation, CommCalculation
+from PyRoute.Calculation.OwnedWorldCalculation import OwnedWorldCalculation
 from PyRoute.Calculation.NoneCalculation import NoneCalculation
 from PyRoute.Calculation.XRouteCalculation import XRouteCalculation
 from PyRoute.StatCalculation import ObjectStatistics
